PycharmProjects/gesture_recognition/requirements.txt
mediapipe==0.10.9
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pyyaml==6.0.1
tqdm==4.66.1
jupyter==1.0.0
ipykernel==6.27.1
imageio==2.33.1
imageio-ffmpeg==0.4.9
python-dotenv==1.0.0